Move signal handlers into natalie.cpp
...and only set up SIGUSR1 and SIGUSR2 handlers if not on macOS.
