+++ src/main.cpp
@@ -1,7 +1,5 @@
 #include "natalie.hpp"
 
-#include <signal.h>
-
 using namespace Natalie;
 
 #ifdef NAT_PRINT_OBJECTS
@@ -43,61 +41,6 @@ extern "C" Object *EVAL(Env *env) {
     }
 }
 
-void sigint_handler(int, siginfo_t *, void *) {
-    const char *msg = "Interrupt\n";
-    auto bytes_written = write(STDOUT_FILENO, msg, strlen(msg));
-    if (bytes_written == -1) abort();
-    exit(128 + SIGINT);
-}
-
-void sigpipe_handler(int, siginfo_t *, void *) {
-    // TODO: do something here?
-}
-
-void gc_signal_handler(int signal, siginfo_t *, void *ucontext) {
-    auto thread = ThreadObject::current();
-    if (!thread || thread->is_main()) return;
-
-    switch (signal) {
-    case SIGUSR1: {
-        auto ctx = thread->get_context();
-        if (!ctx) {
-            char msg[] = "Fatal: Could not get pointer for thread context.";
-            assert(::write(STDERR_FILENO, msg, sizeof(msg)) != -1);
-            abort();
-        }
-        memcpy(ctx, ucontext, sizeof(ucontext_t));
-        thread->set_suspend_status(ThreadObject::SuspendStatus::Suspended);
-#ifdef NAT_DEBUG_THREADS
-        char msg[] = "THREAD DEBUG: Thread suspended\n";
-        assert(::write(STDERR_FILENO, msg, sizeof(msg)) != -1);
-#endif
-        pause();
-        thread->set_suspend_status(ThreadObject::SuspendStatus::Running);
-        break;
-    }
-    case SIGUSR2:
-#ifdef NAT_DEBUG_THREADS
-        char msg2[] = "THREAD DEBUG: Thread resumed\n";
-        assert(::write(STDERR_FILENO, msg2, sizeof(msg2)) != -1);
-#endif
-        thread->set_suspend_status(ThreadObject::SuspendStatus::Running);
-        break;
-    }
-}
-
-void trap_signal(int signal, void (*handler)(int, siginfo_t *, void *)) {
-    struct sigaction sa;
-    sa.sa_sigaction = handler;
-    sigemptyset(&sa.sa_mask);
-    sa.sa_flags = SA_SIGINFO;
-
-    if (sigaction(signal, &sa, nullptr) == -1) {
-        printf("Failed to trap %d\n", signal);
-        exit(1);
-    }
-}
-
 int main(int argc, char *argv[]) {
 #ifdef NAT_NATIVE_PROFILER
     NativeProfiler::enable();
@@ -110,8 +53,10 @@ int main(int argc, char *argv[]) {
 
     trap_signal(SIGINT, sigint_handler);
     trap_signal(SIGPIPE, sigpipe_handler);
+#if !defined(__APPLE__)
     trap_signal(SIGUSR1, gc_signal_handler);
     trap_signal(SIGUSR2, gc_signal_handler);
+#endif
 
 #ifndef NAT_GC_DISABLE
     Heap::the().gc_enable();
